Fixes in style/display, added references, added links and contact for hiring
In title slide, the author name was hidden behind the background image.
I would have liked to get the text area back in the slide (it is only editable
in the master slide) but did not find how to. I ended up moving the background
image to the master slide instead.

I have changed the color of links from an hard to read orange to a dark blue
(blue2) but did not find how to apply the change to all hyperlinks at once.
I had to apply it manually to all links, and all new link still get the orange
color.